requirements.txt
discord.py
better_exceptions
python-dateutil
requests
humanize
